+++ server/types/index.ts
@@ -11,3 +11,5 @@ export * from "./post";
 export * from "./series";
 export * from "./snippet";
 export * from "./tag";
+export * from "./user";
+export * from "./visitor";
